+++ src/main/kotlin/cacophony/controlflow/functions/FunctionHandler.kt
@@ -6,15 +6,17 @@ import cacophony.controlflow.Variable
 import cacophony.controlflow.VariableAllocation
 import cacophony.semantic.syntaxtree.Definition
 
-interface FunctionHandler {
-    fun getFunctionDeclaration(): Definition.FunctionDeclaration
-
+interface FunctionCallHandler {
     fun generateCallFrom(
         callerFunction: FunctionHandler,
         arguments: List<CFGNode>,
         result: Register?,
         respectStackAlignment: Boolean = true,
     ): List<CFGNode>
+}
+
+interface FunctionHandler : FunctionCallHandler {
+    fun getFunctionDeclaration(): Definition.FunctionDeclaration
 
     fun generateVariableAccess(variable: Variable): CFGNode.LValue
