+++ src/main/kotlin/cacophony/controlflow/functions/FunctionHandlerImpl.kt
@@ -9,12 +9,12 @@ import cacophony.semantic.syntaxtree.Definition.FunctionDeclaration
 import kotlin.math.max
 
 class FunctionHandlerImpl(
-    val function: FunctionDeclaration,
+    function: FunctionDeclaration,
     private val analyzedFunction: AnalyzedFunction,
     // List of parents' handlers ordered from immediate parent.
     private val ancestorFunctionHandlers: List<FunctionHandler>,
     callConvention: CallConvention,
-) : FunctionHandler {
+) : FunctionHandler, FunctionCallHandlerImpl(function, ancestorFunctionHandlers) {
     private val staticLink: Variable.AuxVariable.StaticLinkVariable = Variable.AuxVariable.StaticLinkVariable()
     private val definitionToVariable =
         (analyzedFunction.variables.map { it.declaration } union function.arguments).associateWith { Variable.SourceVariable(it) }
@@ -64,25 +64,6 @@ class FunctionHandlerImpl(
     }
 
     override fun getFunctionDeclaration(): FunctionDeclaration = function
-
-    // Wrapper for generateCall that additionally fills staticLink to parent function.
-    // Since staticLink is not property of node itself, but rather of its children,
-    // if caller is immediate parent, we have to fetch RBP instead.
-    override fun generateCallFrom(
-        callerFunction: FunctionHandler,
-        arguments: List<CFGNode>,
-        result: Register?,
-        respectStackAlignment: Boolean,
-    ): List<CFGNode> {
-        val staticLinkVar =
-            if (ancestorFunctionHandlers.isEmpty() || callerFunction === ancestorFunctionHandlers.first()) {
-                RegisterUse(Register.FixedRegister(HardwareRegister.RBP))
-            } else {
-                callerFunction.generateAccessToFramePointer(ancestorFunctionHandlers.first().getFunctionDeclaration())
-            }
-
-        return generateCall(function, arguments + mutableListOf(staticLinkVar), result, respectStackAlignment)
-    }
 
     private fun traverseStaticLink(depth: Int): CFGNode =
         if (depth == 0) {
